+++ sdl/bitarray.h
@@ -25,51 +25,51 @@ class TEBitArray: public IEObject  {
 
 public:
   TEBitArray();
-  TEBitArray( const TEBitArray& arr);
-  TEBitArray(int size);
+  TEBitArray(const TEBitArray& arr);
+  TEBitArray(uint32_t size);
   // if own is true, data [created with new!] will be deleted automatically 
-  TEBitArray(unsigned char* data, size_t size, bool own);
+  TEBitArray(unsigned char* data, uint32_t size, bool own);
   virtual ~TEBitArray();
   void Clear();
-  void SetSize(int newSize);
-  inline int Count()    const {  return (int)FCount;  }
+  void SetSize(uint32_t newSize);
+  inline size_t Count() const {  return FCount;  }
   inline bool IsEmpty() const {  return FCount == 0;  }
-  inline bool operator [] (int index) const  {
-    int intIndex = index/8;
-    int bitIndex = 1 << index%8;
+  inline bool operator [] (size_t index) const  {
+    size_t intIndex = index/8;
+    size_t bitIndex = 1 << index%8;
 #ifdef _OLX_DEBUG
     TIndexOutOfRangeException::ValidateRange(__OlxSourceInfo, intIndex, 0, FIntCount);
 #endif
     return (FData[intIndex] & bitIndex) != 0;
   }
-  bool Get(int index) const  {
-    int intIndex = index/8;
-    int bitIndex = 1 << index%8;
+  bool Get(size_t index) const  {
+    size_t intIndex = index/8;
+    size_t bitIndex = 1 << index%8;
 #ifdef _OLX_DEBUG
     TIndexOutOfRangeException::ValidateRange(__OlxSourceInfo, intIndex, 0, FIntCount);
 #endif
     return (FData[intIndex] & bitIndex) != 0;
   }
-  inline void Set(int index, bool v)  {
-    int intIndex = index/8;
-    int bitIndex = 1 << index%8;
+  inline void Set(size_t index, bool v)  {
+    size_t intIndex = index/8;
+    size_t bitIndex = 1 << index%8;
 #ifdef _OLX_DEBUG
     TIndexOutOfRangeException::ValidateRange(__OlxSourceInfo, intIndex, 0, FIntCount);
 #endif
     if( !v )  FData[intIndex] &= ~bitIndex;
     else      FData[intIndex] |= bitIndex;
   }
-  inline void SetTrue(int index)   {  
-    int intIndex = index/8;
-    int bitIndex = 1 << index%8;
+  inline void SetTrue(size_t index)   {  
+    size_t intIndex = index/8;
+    size_t bitIndex = 1 << index%8;
 #ifdef _OLX_DEBUG
     TIndexOutOfRangeException::ValidateRange(__OlxSourceInfo, intIndex, 0, FIntCount);
 #endif
     FData[intIndex] |= bitIndex;
   }
-  inline void SetFalse(int index)  {  
-    int intIndex = index/8;
-    int bitIndex = 1 << index%8;
+  inline void SetFalse(size_t index)  {  
+    size_t intIndex = index/8;
+    size_t bitIndex = 1 << index%8;
 #ifdef _OLX_DEBUG
     TIndexOutOfRangeException::ValidateRange(__OlxSourceInfo, intIndex, 0, FIntCount);
 #endif
@@ -77,12 +77,12 @@ public:
   }
   void SetAll(bool v);
   inline const unsigned char* GetData() const {  return FData;  }
-  inline int CharCount()                const {  return FCharCount;  }
+  inline size_t CharCount() const {  return FCharCount;  }
 
   void operator << (IInputStream& in);
   void operator >> (IOutputStream& out) const;
   
-  TEBitArray& operator = (const TEBitArray& arr );
+  TEBitArray& operator = (const TEBitArray& arr);
   bool operator == (const TEBitArray& arr )  const;
   int Compare(const TEBitArray& arr )  const;
 
@@ -90,7 +90,7 @@ public:
   void FromHexString(const olxstr& str);
 
   virtual TIString ToString() const;
-  olxstr FormatString( short bitsInSegment );
+  olxstr FormatString(uint16_t bitsInSegment);
 };
 
 EndEsdlNamespace()
